pyproject.toml
[tool.poetry]
name = "opensea-random-nft-pfp-scrapper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
selenium = "^4.17.2"
undetected-chromedriver = "^3.5.4"
webdriver-manager = "^4.0.1"
selenium-stealth = "^1.0.6"
werkzeug = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
